Typveränderung der Variablen X und Y der jeweiligen Models, von "double" auf "float". Die ProgressBar kann man nun nach belieben verschieben (X und Y) oder von der Höhe oder Breite verändern (Height und Width). Die Setter Methode "setPatienten" vom Model "Krankenwagen" wurde angepasst, damit die Liste niemals "Null" werden kann und die Kapazität des Krankenwagens berücksichtigt wird. Entfernung von unnötigen Imports. Anpassung einiger Fragmente, damit informationen übermittelt werden können (z.B.: WarteraumFragment kann jetzt die Anzahl der Patienten bei Bedarf von außen erhalten). Das Layout "fragment_progressbar.xml" wurde angepasst, damit die ProgressBar von der X und Y Position sowie der Höhe und Breite geändert werden kann. Es wurde ein neues Fragment "KrankenwagenFragment" hinzugefügt, welches nur den Krankenwagen auf der Fahrstrecke darstellen soll. Die Logik befindet sich im "KrankenwagenController". Das benötigte Layout "fragment_krankenwagen.xml" wurde sporadisch angelegt (sollte auf jeden Fall durch ein besseres Bild ersetzt werden).

+++ app/src/main/java/de/idlepolicetycoon/project/model/Arzt.java
@@ -9,7 +9,7 @@ public class Arzt extends Mitarbeiter {
     private double geschwindigkeit;
     private double heilung;
 
-    public Arzt(int image, double x, double y, long upgradekosten, int level) {
+    public Arzt(int image, float x, float y, long upgradekosten, int level) {
         super(x, y, image, upgradekosten, level);
     }
